Expand discovery and rankings experience

+++ backend/src/Rewndly.Application/Modules/Admin/AdminContracts.cs
@@ -63,7 +63,7 @@ public sealed record AdminReviewResponse(
     string MediaType,
     int TmdbId,
     string MediaTitle,
-    int? RatingSnapshot,
+    decimal? RatingSnapshot,
     string Title,
     bool ContainsSpoilers,
     string Visibility,
